requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
Appium-Python-Client==1.3.0
atomicwrites==1.4.0
attrs==21.3.0
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
docopt==0.6.2
eyes-common==4.25.3
eyes-core==4.25.3
eyes-selenium==4.25.3
idna==3.3
iniconfig==1.1.1
lxml==4.7.1
mysql-connector==2.2.9
packaging==21.3
Pillow==8.4.0
pipreqs==0.4.11
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
requests==2.26.0
selenium==3.141.0
six==1.16.0
smart-assertions==1.0.2
structlog==20.1.0
tinycss2==1.0.2
toml==0.10.2
transitions==0.8.3
ua-parser==0.10.0
urllib3==1.26.7
webdriver-manager==3.5.2
webencodings==0.5.1
yarg==0.1.9
